requirements.txt
# Vector Database
chromadb>=0.4.22

# Embedding - Text2Vec for Chinese
text2vec>=1.1.5

# Embedding - Sentence Transformers for GTE models
sentence-transformers>=2.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
python-dotenv>=1.0.0
